add new features (login,register,filters) and upgrade views

+++ Controllers/ExpensesController.cs
@@ -1,16 +1,21 @@
+using System.Diagnostics;
+using System.Security.Claims;
 using FluentValidation;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
 using PersonalExpenses.Entities;
+using PersonalExpenses.Helpers;
 using PersonalExpenses.Services;
 using PersonalExpenses.ViewModels;
 
 namespace PersonalExpenses.Controllers;
 
+[Authorize]
 public class ExpensesController : Controller
 {
     private readonly IExpenseService _expenseService;
     private readonly ICategoryService _categoryService;
-
     private readonly IValidator<CreateExpenseModel> _validator;
 
     public ExpensesController(IExpenseService expenseService, ICategoryService categoryService, IValidator<CreateExpenseModel> validator)
@@ -21,11 +26,42 @@ public class ExpensesController : Controller
     }
 
     [HttpGet]
-    // GET: Expenses
-    public async Task<IActionResult> Index()
+    // GET: Expenses for the logged-in user
+    public async Task<IActionResult> Index([FromQuery(Name = "Filter")] ExpenseFilterViewModel filterForm)
     {
-        var expenses = await _expenseService.GetAllAsync();
-        return View(expenses);
+
+        IEnumerable<Expense> expenses;
+        filterForm.Categories = (await _categoryService.GetAllAsync())
+            .Select(c => new SelectListItem { Value = c.Id.ToString(), Text = c.Name }).ToList();
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+        // Set up required filters
+        var filters = SetUpFilters(filterForm, userId);
+        var totalThisMonth = await _expenseService.GetTotalForCurrentMonthAsync(userId!);
+        expenses = filters.Count == 0 ? await _expenseService.GetAllByUserAsync(userId!) : await _expenseService.GetForUserAsyncWith(userId!, filters);
+
+        return View(new ExpenseListViewModel()
+        {
+            Expenses = expenses.ToList(),
+            Filter = filterForm,
+            TotalThisMonth = totalThisMonth,
+
+
+        });
+    }
+
+    private List<IExpenseFilter> SetUpFilters(ExpenseFilterViewModel filter, string? userId)
+    {
+        List<IExpenseFilter> filters = [];
+        if (filter.FromDate.HasValue || filter.ToDate.HasValue)
+        {
+            filters.Add(new ExpenseBetweenDatesFilter(filter.FromDate, filter.ToDate));
+        }
+        if (filter.CategoryId.HasValue)
+        {
+            filters.Add(new ExpenseCategoryFilter(filter.CategoryId.Value));
+        }
+        return filters;
     }
 
     [HttpGet("Expenses/Create")]
@@ -61,16 +97,21 @@ public class ExpensesController : Controller
             return View(createExpenseModel);
         }
 
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+        if (string.IsNullOrEmpty(userId)) return BadRequest("User ID is required.");
+
         var expense = new Expense()
         {
             Amount = createExpenseModel.Expense,
             Description = createExpenseModel.Description,
             Currency = createExpenseModel.Currency,
             CategoryId = createExpenseModel.CategoryId,
-            Date = createExpenseModel.Date
+            Date = createExpenseModel.Date,
+            UserId = userId
         };
         await _expenseService.CreateAsync(expense);
-
+        TempData["SuccessMessage"] = "Gasto agregado con éxito.";
         return RedirectToAction(nameof(Index));
     }
 
@@ -79,6 +120,8 @@ public class ExpensesController : Controller
     [ValidateAntiForgeryToken]
     public async Task<IActionResult> DeleteConfirmed(int id)
     {
+        var expense = await _expenseService.GetByIdAsync(id);
+        if (expense == null || expense.UserId != User.FindFirstValue(ClaimTypes.NameIdentifier)) return NotFound();
         await _expenseService.DeleteAsync(id);
         return RedirectToAction(nameof(Index));
     }
@@ -87,8 +130,9 @@ public class ExpensesController : Controller
     public async Task<IActionResult> Edit(int id)
     {
         var expense = await _expenseService.GetByIdAsync(id);
-        if (expense == null) return NotFound();
         var categories = await _categoryService.GetAllAsync();
+
+        if (expense == null || expense.UserId != User.FindFirstValue(ClaimTypes.NameIdentifier)) return RedirectToAction("Error404", "Error");
 
         var editExpenseModel = new EditExpenseModel()
         {
@@ -98,7 +142,6 @@ public class ExpensesController : Controller
             Date = expense.Date,
             CategoryId = expense.CategoryId,
             Expense = expense.Amount,
-            Id = expense.Id
         };
         return View(editExpenseModel);
     }
@@ -108,21 +151,21 @@ public class ExpensesController : Controller
     [ValidateAntiForgeryToken]
     public async Task<IActionResult> Edit(int id, EditExpenseModel editExpenseModel)
     {
-        if (id != editExpenseModel.Id) return BadRequest();
+        var expense = await _expenseService.GetByIdAsync(id);
+
+        if (expense == null || expense.UserId != User.FindFirstValue(ClaimTypes.NameIdentifier)) return NotFound();
 
         if (!ModelState.IsValid)
         {
             ViewBag.Categories = await _categoryService.GetAllAsync();
             return View(editExpenseModel);
         }
-        var expense = new Expense()
-        {
-            Amount = editExpenseModel.Expense,
-            Description = editExpenseModel.Description,
-            Currency = editExpenseModel.Currency,
-            CategoryId = editExpenseModel.CategoryId,
-            Date = editExpenseModel.Date
-        };
+
+        expense.Amount = editExpenseModel.Expense;
+        expense.Description = editExpenseModel.Description;
+        expense.Currency = editExpenseModel.Currency;
+        expense.CategoryId = editExpenseModel.CategoryId;
+        expense.Date = editExpenseModel.Date;
 
         await _expenseService.UpdateAsync(expense);
         return RedirectToAction(nameof(Index));
@@ -133,16 +176,16 @@ public class ExpensesController : Controller
     public async Task<IActionResult> Details(int id)
     {
         var expense = await _expenseService.GetByIdAsync(id);
-        if (expense == null) return NotFound();
+        if (expense == null || expense.UserId != User.FindFirstValue(ClaimTypes.NameIdentifier)) return NotFound();
         var detailsModel = new ExpenseDetailsModel()
         {
             Id = expense.Id,
             Expense = expense.Amount,
             Date = expense.Date,
+            Description = expense.Description,
             CategoryName = expense.Category?.Name ?? string.Empty
         };
         return View(detailsModel);
     }
-
 
 }
